requirements.txt
python-telegram-bot==21.0.1
google-generativeai==0.8.3
supabase==2.7.4
python-dotenv==1.0.1
fuzzywuzzy==0.18.0
python-levenshtein==0.25.1
asyncio==3.4.3
uvloop==0.19.0